+++ ExperPort/Analysis/NL2AFC_MIX/Subplot_summry_070227.m
@@ -1,5 +1,5 @@
 figure;
-set(gcf,'Position',[112 465 1087 633]);
+set(double(gcf),'Position',[112 465 1087 633]);
 n_epoch = [3 3 3 2 2 2]; 
 pair1 = 'EB/Prp'; pair2 = 'Hex/CA';
 load('C:\Home\RatExper\SoloData\Data\NL_Analysis\nl2afc_mix2\data_summ_070226a');
@@ -9,7 +9,7 @@ for i = 1:6,
     subplot(2,3,i); hold on; 
     plot((1:n_epoch(i)),cell2mat(mix2afc_ratname(2:n_epoch(i)+1,2)),'o-b', 'MarkerSize', 10);
     plot((1:n_epoch(i)),cell2mat(mix2afc_ratname(2:n_epoch(i)+1,3)),'*-r', 'MarkerSize', 12);
-    set(gca,'XTick',(1:1:n_epoch(i)),'XTickLabel',{mix2afc_ratname{2:n_epoch(i)+1,1}},...
+    set(double(gca),'XTick',(1:1:n_epoch(i)),'XTickLabel',{mix2afc_ratname{2:n_epoch(i)+1,1}},...
         'YLim',[30 100],'YGrid','on'); xlabel('Pair1 Frequency');
     legend(pair1, pair2);
 end;
